[numerics] clean up tools tests
