+++ numerics/src/tools/test/test_pinv.c
